+++ inc/admin/auth.php
@@ -244,7 +244,7 @@ function wolf_core_wrong_theme() {
 		'speaker',
 		'nu',
 		'galatea',
-		'stork',
+		'spectacle',
 	);
 
 	return ( ! in_array( esc_attr( sanitize_title_with_dashes( get_template() ) ), $ok, true ) );