+++ inc/admin/auth.php
@@ -238,13 +238,12 @@ function wolf_core_wrong_theme() {
 		'oglin',
 		'staaw',
 		'bronze',
-		'hazal',
-		'craftz',
 		'swingster',
 		'speaker',
 		'nu',
 		'galatea',
 		'spectacle',
+		'effigy',
 	);
 
 	return ( ! in_array( esc_attr( sanitize_title_with_dashes( get_template() ) ), $ok, true ) );