+++ inc/admin/classes/class-metaboxes.php
@@ -203,7 +203,7 @@ if ( ! class_exists( 'Wolf_Metaboxes' ) ) {
 						<span class="description">' . wolf_core_kses( $desc ) . '</span>';
 
 			// datepicker
-					} elseif ( $field['type'] == 'datepicker' ) {
+			} elseif ( $field['type'] == 'datepicker' ) {
 						echo '<input type="text" class="wd-metabox-datepicker" name="' . $field['id'] . '" id="' . $field['id'] . '" value="' . $meta . '" size="30">
 						<br><span class="description">' . $desc . '</span>';