+++ plugins/elements/album-disc.php
@@ -70,10 +70,13 @@ function wolf_core_album_disc( $atts ) {
 	$output .= wolf_core_element_aos_animation_data_attr( $atts );
 	$output .= '>';
 
-	if ( is_array( $link ) && isset( $link['url'] ) ) {
+	if ( is_array( $link ) && isset( $link['url'] ) && '' !== $link['url'] ) {
 		$output .= '<a rel="' . esc_attr( $link['rel'] ) . '" class="wolf-core-album-disc-link-mask"';
 		$output .= ' target="' . esc_attr( $link['target'] ) . '"';
 		$output .= ' href="' . esc_url( $link['url'] ) . '" title="' . esc_attr( $link['title'] ) . '"></a>';
+	} else {
+		$output .= '<a class="wolf-core-album-disc-link-mask"';
+		$output .= ' href="' . get_the_permalink() . '" title="' . get_the_title() . '"></a>';
 	}
 
 	$output .= '<div class="wolf-core-album-disc-cover-container">';